Move validation of configuration into the main function
If we do the validation after, we wait much longer to know our
configuration is not working. By having it at the start, we can then
rely on it being present, and failing early if the configuration is bad.

It also makes more implicit that `getIussueSummary` really only gets the
issues summary and does not actually do validation on the configuration